Modularize framebus
