+++ package.json
@@ -7,7 +7,7 @@
   },
   "homepage": "https://github.com/braintree/framebus",
   "version": "3.0.2",
-  "main": "dist/framebus.js",
+  "main": "dist/index.js",
   "files": [
     "dist/"
   ],
