+++ spec/unit/attach.spec.ts
@@ -1,22 +1,23 @@
-import bus = require("../../src/lib/framebus");
+import { attach, detach } from "../../src/lib/attach";
+import onmessage from "../../src/lib/message";
 
 describe("_attach", function () {
   let addEventSpy: jest.SpyInstance;
 
   beforeEach(function () {
-    bus._detach();
+    detach();
     addEventSpy = jest.spyOn(window, "addEventListener").mockImplementation();
   });
 
   it("should add listener to scope", function () {
-    bus._attach();
+    attach();
 
-    expect(addEventSpy).toBeCalledWith("message", bus._onmessage, false);
+    expect(addEventSpy).toBeCalledWith("message", onmessage, false);
   });
 
   it("should only add listener to scope once", function () {
-    bus._attach();
-    bus._attach();
+    attach();
+    attach();
 
     expect(addEventSpy).toBeCalledTimes(1);
   });
