+++ spec/unit/emit.spec.ts
@@ -1,21 +1,14 @@
-import bus = require("../../src/lib/framebus");
+import { attach } from "../../src/lib/attach";
+import bus = require("../../src/");
 
-describe("publish", function () {
+describe("emit", function () {
   beforeEach(function () {
-    bus._attach();
-  });
-
-  it("should be directly usable", function () {
-    const publish = bus.publish;
-
-    expect(function () {
-      publish("event", { data: "data" });
-    }).not.toThrowError();
+    attach();
   });
 
   it("should return false if event is not a string", function () {
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    const actual = bus.publish({} as any, { data: "data" });
+    const actual = bus.emit({} as any, { data: "data" });
 
     expect(actual).toBe(false);
   });
@@ -24,7 +17,7 @@ describe("publish", function () {
     const actual = bus
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
       .target({ origin: "object" } as any)
-      .publish("event", { data: "data" });
+      .emit("event", { data: "data" });
 
     expect(actual).toBe(false);
   });
@@ -32,15 +25,13 @@ describe("publish", function () {
   it("should return true if origin and event are strings", function () {
     const actual = bus
       .target("https://example.com")
-      .publish("event", { data: "data" });
+      .emit("event", { data: "data" });
 
     expect(actual).toBe(true);
   });
 
   it("can pass a reply function without passing data", function () {
-    const actual = bus
-      .target("https://example.com")
-      .publish("event", jest.fn());
+    const actual = bus.target("https://example.com").emit("event", jest.fn());
 
     expect(actual).toBe(true);
   });
