+++ spec/unit/packagePayload.spec.ts
@@ -1,12 +1,12 @@
-import bus = require("../../src/lib/framebus");
+import packagePayload from "../../src/lib/package-payload";
 
 const messagePrefix = "/*framebus*/";
 
-describe("_packagePayload", function () {
+describe("packagePayload", function () {
   it("should add event to payload", function () {
     const expected = "event name";
 
-    const result = bus._packagePayload.call(bus, expected, "*", {});
+    const result = packagePayload(expected, "*", {});
     const actual = JSON.parse(result.replace(messagePrefix, "")).event;
 
     expect(actual).toBe(expected);
@@ -15,14 +15,14 @@ describe("_packagePayload", function () {
   it("should add data to payload", function () {
     const expected = { some: "data" };
 
-    const result = bus._packagePayload.call(bus, "event", "*", expected);
+    const result = packagePayload("event", "*", expected);
     const actual = JSON.parse(result.replace(messagePrefix, ""));
 
     expect(actual.eventData).toEqual(expected);
   });
 
   it("should add reply to payload if provided", function () {
-    const result = bus._packagePayload.call(bus, "event", "*", {}, jest.fn());
+    const result = packagePayload("event", "*", {}, jest.fn());
     const actual = JSON.parse(result.replace(messagePrefix, ""));
 
     expect(typeof actual.reply).toBe("string");
@@ -40,7 +40,7 @@ describe("_packagePayload", function () {
     });
 
     const fn = function (): void {
-      bus._packagePayload("event", "*", payload);
+      packagePayload("event", "*", payload);
     };
 
     expect(fn).toThrowError("Could not stringify event: ");
