+++ src/index.ts
@@ -1,0 +1,8 @@
+import { attach } from "./lib/attach";
+import Framebus from "./framebus";
+
+const bus = new Framebus();
+
+attach();
+
+export = bus;
