+++ src/lib/is-not-string.ts
@@ -1,0 +1,3 @@
+export default function isntString(str: string): boolean {
+  return typeof str !== "string";
+}
